Additional Refactoring
